Fix typo in goodbye component and update hello component
Fix typo in the goodbye component by changing "Good bye1" to "Good bye". Also, update the hello component by removing the trailing "122" in the div with id "hello-world". This ensures that the hello component displays the correct text.

Update the index.html file

Update the index.html file to reflect the changes made in the hello and goodbye components. Replace the "Hello world!122" text in the first hello-world div with just "Hello world!". Similarly, replace the "Good bye1" text with "Good bye" in the second div. These changes ensure that the index.html file displays the correct content.